+++ core/src/Kokkos_Core_fwd.hpp
@@ -278,7 +278,8 @@ namespace Kokkos {
 template <class DataType, class... Properties>
 class View;
 #ifndef KOKKOS_ENABLE_IMPL_VIEW_LEGACY
-template <class DataType, class... Properties>
+template <class ElementType, class Extents, class LayoutPolicy,
+          class AccessorPolicy>
 class BasicView;
 #endif
