+++ core/src/impl/Kokkos_ViewMapping.hpp
@@ -41,27 +41,6 @@
 //----------------------------------------------------------------------------
 
 namespace Kokkos {
-
-struct ALL_t {
-  KOKKOS_INLINE_FUNCTION
-  constexpr const ALL_t& operator()() const { return *this; }
-
-  KOKKOS_INLINE_FUNCTION
-  constexpr bool operator==(const ALL_t&) const { return true; }
-};
-
-#ifdef KOKKOS_ENABLE_DEPRECATED_CODE_4
-namespace Impl {
-// TODO This alias declaration forces us to fully qualify ALL_t inside the
-// Kokkos::Impl namespace to avoid deprecation warnings. Replace the
-// fully-qualified name when we remove Kokkos::Impl::ALL_t.
-using ALL_t KOKKOS_DEPRECATED_WITH_COMMENT("Use Kokkos::ALL_t instead!") =
-    Kokkos::ALL_t;
-}  // namespace Impl
-#endif
-}  // namespace Kokkos
-
-namespace Kokkos {
 namespace Impl {
 
 template <class T>
